refactor #36 중복코드 최소화
