+++ db.json
@@ -867,9 +867,9 @@
     {
       "duration": "2025-04",
       "categoryId": 1,
-      "sumAmount": 252400,
+      "sumAmount": 262400,
       "createdAt": "2025-04-01T00:00:00Z",
-      "updatedAt": "2025-04-10T07:49:22.209Z",
+      "updatedAt": "2025-04-11T00:00:00.000Z",
       "userId": "7471",
       "id": "20d4"
     },
@@ -885,9 +885,9 @@
     {
       "duration": "2025-04",
       "categoryId": 12,
-      "sumAmount": 103500,
+      "sumAmount": 114500,
       "createdAt": "2025-04-01T00:00:00Z",
-      "updatedAt": "2025-04-01T00:00:00Z",
+      "updatedAt": "2025-04-03T00:00:00.000Z",
       "userId": "7471",
       "id": "9a03"
     },
@@ -1704,9 +1704,9 @@
     {
       "duration": "2025-04",
       "categoryId": 2,
-      "sumAmount": 90600,
+      "sumAmount": 91600,
       "createdAt": "2025-04-01T00:00:00Z",
-      "updatedAt": "2025-04-01T00:00:00Z",
+      "updatedAt": "2025-04-24T00:00:00.000Z",
       "userId": "7471",
       "id": "c100"
     },
@@ -12252,7 +12252,7 @@
       "vendor": "회사",
       "transactionType": "expense",
       "createdAt": "2025-04-01T00:00:00.000Z",
-      "updatedAt": "2025-04-10T05:05:26.231Z",
+      "updatedAt": "2025-04-10T13:17:15.438Z",
       "userId": "7471",
       "id": "2bbc",
       "date": ""
@@ -15288,10 +15288,11 @@
       "paymentMethod": "카드",
       "vendor": "마트",
       "transactionType": "expense",
-      "createdAt": "2025-04-02T00:00:00Z",
-      "updatedAt": "2025-04-02T00:00:00Z",
+      "createdAt": "2025-04-02T00:00:00.000Z",
+      "updatedAt": "2025-04-10T13:15:55.756Z",
       "userId": "7471",
-      "id": "a7e8"
+      "id": "a7e8",
+      "date": ""
     },
     {
       "categoryId": 16,
@@ -15368,14 +15369,15 @@
     {
       "categoryId": 4,
       "amount": 81700,
-      "memo": "Dummy expense",
+      "memo": "Dummy",
       "paymentMethod": "카드",
       "vendor": "기타",
       "transactionType": "expense",
-      "createdAt": "2025-04-02T00:00:00Z",
-      "updatedAt": "2025-04-02T00:00:00Z",
+      "createdAt": "2025-04-02T00:00:00.000Z",
+      "updatedAt": "2025-04-10T13:16:03.179Z",
       "userId": "7471",
-      "id": "0713"
+      "id": "0713",
+      "date": ""
     },
     {
       "categoryId": 13,
@@ -15668,15 +15670,99 @@
     {
       "id": "0473",
       "categoryId": 1,
-      "memo": "맛집인듯 ㅇㅇ",
-      "paymentMethod": "card",
+      "memo": "맛집인듯",
+      "paymentMethod": "카드",
       "amount": 20000,
       "vendor": "중국집",
-      "transactionType": "",
+      "transactionType": "expense",
       "createdAt": "2025-04-01T00:00:00.000Z",
-      "updatedAt": "2025-04-10T07:49:22.209Z",
+      "updatedAt": "2025-04-10T13:26:08.494Z",
       "userId": "7471",
       "date": ""
+    },
+    {
+      "id": "01a3",
+      "categoryId": 2,
+      "memo": "굿",
+      "paymentMethod": "카드",
+      "amount": "1000",
+      "vendor": "버스",
+      "transactionType": "expense",
+      "createdAt": "2025-04-24T00:00:00.000Z",
+      "updatedAt": "2025-04-24T00:00:00.000Z",
+      "userId": "7471"
+    },
+    {
+      "id": "9004",
+      "categoryId": 1,
+      "memo": "ㅇㅇ",
+      "paymentMethod": "카드",
+      "amount": "10000",
+      "vendor": "ㅇ",
+      "transactionType": "expense",
+      "createdAt": "2025-04-11T00:00:00.000Z",
+      "updatedAt": "2025-04-11T00:00:00.000Z",
+      "userId": "7471"
+    },
+    {
+      "id": "a7f6",
+      "categoryId": 12,
+      "memo": "ㅗㅗㅗㅗ",
+      "paymentMethod": "카드",
+      "amount": "10000",
+      "vendor": "ㅗ",
+      "transactionType": "expense",
+      "createdAt": "2025-10-23T00:00:00.000Z",
+      "updatedAt": "2025-10-23T00:00:00.000Z",
+      "userId": "7471"
+    },
+    {
+      "id": "d572",
+      "categoryId": 12,
+      "memo": "ㅇㅇ",
+      "paymentMethod": "카드",
+      "amount": "110000",
+      "vendor": "술잡",
+      "transactionType": "expense",
+      "createdAt": "2025-04-11T00:00:00.000Z",
+      "updatedAt": "2025-04-11T00:00:00.000Z",
+      "userId": "7471"
+    },
+    {
+      "id": "23af",
+      "categoryId": 12,
+      "memo": "ㅇㅇ",
+      "paymentMethod": "카드",
+      "amount": "10000",
+      "vendor": "ㅇ",
+      "transactionType": "expense",
+      "createdAt": "2025-04-21T00:00:00.000Z",
+      "updatedAt": "2025-04-21T00:00:00.000Z",
+      "userId": "7471"
+    },
+    {
+      "id": "3aad",
+      "categoryId": 12,
+      "memo": "ㅂㅈㄷ",
+      "paymentMethod": "카드",
+      "amount": "10000",
+      "vendor": "ㄷㅂㄷ",
+      "transactionType": "expense",
+      "createdAt": "2025-04-16T00:00:00.000Z",
+      "updatedAt": "2025-04-16T00:00:00.000Z",
+      "userId": "7471"
+    },
+    {
+      "id": "bf2f",
+      "categoryId": 12,
+      "memo": "ㅈㅂㄷ",
+      "paymentMethod": "카드",
+      "amount": 1000,
+      "vendor": "ㅉ",
+      "transactionType": "expense",
+      "createdAt": "2025-04-03T00:00:00.000Z",
+      "updatedAt": "2025-04-03T00:00:00.000Z",
+      "userId": "7471"
     }
   ],
   "category": [
